Update register screen to be observer
