+++ app/navigation/root-navigator.tsx
@@ -2,9 +2,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import type React from 'react';
 
-import LoginScreen from '~/views/login-screen/login-screen';
-import RegisterScreen from '~/views/register';
-
+import { LoginScreen, RegisterScreen } from '~/views';
 import type { RootParamList } from './root-params';
 import { TabNavigator } from './tab-navigator';
