+++ app/views/index.ts
@@ -1,1 +1,2 @@
 export * from './login-screen';
+export * from './register-screen';
