+++ app/views/register-screen/index.ts
@@ -1,0 +1,2 @@
+export * from './register-screen';
+export * from './register-screen.props';
