+++ app/views/register-screen/register-screen.props.ts
@@ -1,0 +1,1 @@
+export interface RegisterScreenProps {}
